fix(lint): Correct type comparisons and idiomatic patterns
terminal/raw.rs:
- Remove unnecessary "as usize" cast on NCCS constant; libc types
  already match on this platform

unicode/bidi.rs:
- Replace .iter().any(|&l| l == N) with .contains(&N) for level checks
- More idiomatic and clearer intent

+++ src/terminal/raw.rs
@@ -244,8 +244,8 @@ mod tests {
     #[test]
     fn test_termios_control_chars() {
         // VMIN and VTIME indices should be valid
-        assert!(libc::VMIN < libc::NCCS as usize);
-        assert!(libc::VTIME < libc::NCCS as usize);
+        assert!(libc::VMIN < libc::NCCS);
+        assert!(libc::VTIME < libc::NCCS);
     }
 
     // ============================================
